+++ esmerald/datastructures/__init__.py
@@ -14,7 +14,7 @@ from .base import (
     URLPath,
 )
 from .file import File
-from .json import JSON, OrJSON
+from .json import JSON
 from .redirect import Redirect
 from .stream import Stream
 from .template import Template
@@ -27,7 +27,6 @@ __all__ = [
     "Headers",
     "JSON",
     "MutableHeaders",
-    "OrJSON",
     "QueryParams",
     "Redirect",
     "ResponseContainer",
