+++ esmerald/datastructures/json.py
@@ -2,7 +2,6 @@ from typing import TYPE_CHECKING, Any, Dict, Optional, Type, Union
 
 from esmerald.datastructures.base import ResponseContainer  # noqa
 from esmerald.responses import JSONResponse  # noqa
-from esmerald.responses import ORJSONResponse  # noqa
 
 if TYPE_CHECKING:
     from esmerald.applications import Esmerald
@@ -33,8 +32,3 @@ class JSON(ResponseContainer[JSONResponse]):
             media_type=media_type,
             background=self.background,
         )
-
-
-class OrJSON(JSON, ResponseContainer[ORJSONResponse]):
-    ...
-
