+++ esmerald/responses/json.py
@@ -25,13 +25,3 @@ class BaseJSONResponse(JSONResponse):
         if isinstance(value, BaseModel):
             return value.dict()
         raise TypeError("unsupported type")
-
-
-class ORJSONResponse(BaseJSONResponse):
-    def render(self, content: Any) -> bytes:
-        assert orjson is not None, "orjson must be installed to use ORJSONResponse"
-        return orjson.dumps(
-            content,
-            default=self.transform,
-            option=OPT_SERIALIZE_NUMPY | OPT_OMIT_MICROSECONDS,
-        )
