+++ tests/test_datastructures_response.py
@@ -1,8 +1,8 @@
 import pytest
 
 from esmerald import Gateway, get, status
-from esmerald.datastructures import JSON, OrJSON
-from esmerald.datastructures.encoders import UJSON
+from esmerald.datastructures import JSON
+from esmerald.datastructures.encoders import UJSON, OrJSON
 from esmerald.testclient import create_client
