+++ tests/test_headers.py
@@ -1,8 +1,10 @@
-from esmerald import Gateway, Header, ORJSONResponse, Param, Response, post
+from pydantic import BaseModel
+
+from esmerald import Gateway, Header, Param, Response, post
 from esmerald.datastructures import ResponseHeader
 from esmerald.enums import MediaType
+from esmerald.responses.encoders import ORJSONResponse
 from esmerald.testclient import create_client
-from pydantic import BaseModel
 
 
 class User(BaseModel):
